IP routing: ipv4 routing on OpenTL, AcceptTL, AuditTL and KeysSharingTL transactions;

+++ src/core/contractors/addresses/BaseAddress.h
@@ -9,7 +9,7 @@ class BaseAddress {
 
 public:
     typedef shared_ptr<BaseAddress> Shared;
-    typedef uint16_t SerializedType;
+    typedef uint8_t SerializedType;
 
     enum AddressType {
         /*
